infotrophic_web_app/requirements.txt
Django==1.8.5
Unidecode==0.4.18
Whoosh==2.7.0
argparse==1.2.1
beautifulsoup4==4.4.0
boto==2.38.0
django-classy-tags==0.6.2
django-cors-headers==1.1.0
django-haystack==2.4.0
django-reversion==1.8.0
django-storages==1.1.8
six==1.10.0
wsgiref==0.1.2
